+++ internal/controller/http/handler.go
@@ -1,11 +1,9 @@
 package http
 
 import (
-	"encoding/json"
 	"net/http"
 
 	"github.com/probuborka/go_final_project/internal/entity"
-	"github.com/probuborka/go_final_project/pkg/logger"
 )
 
 type handler struct {
@@ -46,15 +44,4 @@ func (h handler) Init() *http.ServeMux {
 	r.HandleFunc("DELETE /api/task", h.deleteTask)
 
 	return r
-}
-
-func response(w http.ResponseWriter, v any, statusCode int) {
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	w.WriteHeader(statusCode)
-	resp, err := json.Marshal(&v)
-	if err != nil {
-		logger.Error(err)
-		return
-	}
-	w.Write(resp)
 }
